requirements.txt
pandas
numpy
PyYAML
opencv-python
Pillow
matplotlib
seaborn
albumentations
streamlit
ultralytics
